Doc generation working
Using `pydoc-markdown` together with the developer's new `novella`
build system, which does _not_ support generating API docs.

https://niklasrosenstein.github.io/pydoc-markdown/getting-started/#static-vs-generated-markdown-files

https://github.com/NiklasRosenstein/pydoc-markdown/discussions/254

So, we need to do some script-y silliness to generate stub Markdown
files for each module in the package. This code is in the new
`splatlog._docs` module, and is a hacky-mess at the moment, but it
does seem to do the trick.